final-pjt-back/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.12
